+++ src/main/java/sophisticated_wolves/item/ItemDogTreat.java
@@ -1,12 +1,17 @@
 package sophisticated_wolves.item;
 
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.MobSpawnType;
 import net.minecraft.world.entity.animal.Wolf;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.gameevent.GameEvent;
+import sophisticated_wolves.core.SWEntities;
 import sophisticated_wolves.core.SWTabs;
 import sophisticated_wolves.api.ISophisticatedWolf;
 import sophisticated_wolves.compatibility.Compatibility;
@@ -34,35 +39,39 @@ public class ItemDogTreat extends Item {
                 entity instanceof Wolf wolf &&
                 !(entity instanceof ISophisticatedWolf)) {
             if (wolf.isTame()) {
-                //TODO
-                SophisticatedWolf sWolf = null;//new SophisticatedWolf(player.getLevel());
+                var level = player.getLevel();
+                var sWolf = (SophisticatedWolf) SWEntities.getSophisticatedWolfType().spawn(
+                        (ServerLevel) level, stack, player,
+                        entity.blockPosition(), MobSpawnType.SPAWN_EGG,
+                        true, true);
+                if (sWolf != null) {
+                    sWolf.copyPosition(wolf);
+                    sWolf.setCustomName(wolf.getCustomName());
+                    sWolf.setCollarColor(wolf.getCollarColor());
+                    sWolf.setTame(true);
+                    sWolf.setOwnerUUID(wolf.getOwnerUUID());
+                    sWolf.setHealth(wolf.getHealth());
+                    sWolf.setWolfSpeciesByBiome();
 
-                sWolf.copyPosition(wolf);
-                sWolf.setCustomName(wolf.getCustomName());
-                sWolf.setCollarColor(wolf.getCollarColor());
-                sWolf.setTame(true);
-                sWolf.setOwnerUUID(wolf.getOwnerUUID());
-                sWolf.setHealth(wolf.getHealth());
-                sWolf.setWolfSpeciesByBiome();
+                    if (Compatibility.IS_WOLF_ARMOR_INSTALLED) {
+                        //TODO
+                        //CompatibilityWolfArmor.copyWolfItems(wolf, sWolf);
+                    }
 
-                if (Compatibility.IS_WOLF_ARMOR_INSTALLED) {
+                    wolf.remove(Entity.RemovalReason.DISCARDED);
+
                     //TODO
-//                    CompatibilityWolfArmor.copyWolfItems(wolf, sWolf);
+                    //player.getLevel().spawnEntity(sWolf);
+
+                    level.gameEvent(player, GameEvent.ENTITY_PLACE, entity.blockPosition());
+
+                    stack.shrink(1);
                 }
-
-                //TODO
-//                wolf.setDead();
-
-                //TODO
-//                player.getLevel().spawnEntity(sWolf);
-                player.getLevel().playSound(player, sWolf, null, null, 1016, 0);
-                stack.shrink(1);
                 return InteractionResult.SUCCESS;
             }
-            return InteractionResult.FAIL;
-        } else {
-            return super.interactLivingEntity(stack, player, entity, hand);
+            return InteractionResult.PASS;
         }
+        return InteractionResult.PASS;
     }
 
 }
